% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{.api_request}
\alias{.api_request}
\title{base request function}
\usage{
.api_request(
  url,
  ...,
  verb = "GET",
  user_agent = getOption("nflversepff.user_agent"),
  retry_delay = 1L,
  retry_times = 3L,
  log_url = url
)
}
\arguments{
\item{url}{Full URL to be scraped.}

\item{...}{arguments to pass to httr request function}

\item{verb}{character: a curl verb supported by httr, e.g. GET, POST, PUT, DELETE
this is passed to \code{httr::RETRY()}. Defaults to \code{"GET"}}

\item{user_agent}{Request user agent, defaults to option \code{nflversepff.user_agent}}

\item{retry_times}{Maximum number of attempts. Defaults to \code{3}.}

\item{log_url}{url for print logging, defaults to \code{url}}
}
\description{
Formats api request to the PFF API with retry
}
