% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{.pff_request}
\alias{.pff_request}
\title{pff request function}
\usage{
.pff_request(endpoint, query, base_url = "https://www.pff.com/api", ...)
}
\arguments{
\item{endpoint}{character: path to access}

\item{query}{a list to be appended as HTTP query parameters (e.g. ?x=x&y=y)}

\item{base_url}{base url, defaults to "https://www.pff.com/api"}

\item{...}{arguments passed to \code{.api_request()}}
}
\description{
Interpolates endpoint and query parameters to the default base URL and then
passes it to \code{.api_request}
}
