% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{pff_teams}
\alias{pff_teams}
\title{Get teams from PFF API}
\usage{
pff_teams(
  season = nflreadr::get_current_season(roster = TRUE),
  league = "nfl",
  ...
)
}
\arguments{
\item{season}{season to retrieve}

\item{league}{one of \code{c('nfl', 'ncaa')}, defaults to nfl}
}
\description{
Collects whatever is available from the "premium" endpoint without authentication
}
