% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_nflverse.R
\name{scrape_players_to_release}
\alias{scrape_players_to_release}
\title{Scrape all PFF players and save to PFF repo release}
\usage{
scrape_players_to_release()
}
\description{
Scrape all PFF players and save to PFF repo release
}
